fix: move labels to separate file
